% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h4decoder95.R
\name{compile_primer}
\alias{compile_primer}
\title{compile_primer}
\usage{
compile_primer(up_seq, down_seq, barcodes)
}
\arguments{
\item{up_seq}{Primer 5' sequence up to the barcode.}

\item{down_seq}{Primer 3' sequence down to the barcode.}

\item{barcodes}{Dataframe of filtered barcodes.}
}
\description{
It generates full sequence of sequencing primers according to your design. It fuses adaptor, barcode and primer sequence of your choice.
}
\examples{
print("Compiling barcodes...")
barcodes = generate_barcode()
filtered_barcodes = filter_barcode(barcodes)
compiled_primer = compile_primer("AAA", "CCC", filtered_barcodes[,1])
print("done.")

}
