% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h4decoder95.R
\name{correct_barcode}
\alias{correct_barcode}
\title{correct_barcode}
\usage{
correct_barcode(index)
}
\arguments{
\item{index}{Vector of barcodes obtained from sequencing.}
}
\description{
It corrects barcodes obtained by sequencing. The capacity of correction of the algorithm is one mismatch maximum.
}
\examples{
print("Correcting index...")
index=c("GACATTGGG","CCTCTACAA","CATATGTGG","ACTCTCGCT","GTTTTAGGA") 
# correct barcodes are GATATTGGG, CCTCTTCAA, CATATGTGG, ACTCTCTCT, GTTTTAGGG 
corrected_index = correct_barcode(index)
print("done.")
print(paste(corrected_index[[3]],"errors have been corrected within the",corrected_index[[2]], "index list provided"))

}
