% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h4decoder95.R
\name{filter_barcode}
\alias{filter_barcode}
\title{filter_barcode}
\usage{
filter_barcode(barcodeslist)
}
\arguments{
\item{barcodeslist}{A list of barcodes to be filtered}
}
\description{
It filters barcodes list according to GC contents. GC content calculation is based on Ringo package from bioconductor.
}
\examples{
print("Filtering barcodes...")
barcodes = generate_barcode()
filtered_barcodes = filter_barcode(barcodes)
print("done.")
print(paste("Only",length(filtered_barcodes$barcodeslist), "barcodes remain. "))

}
