% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h4decoder95.R
\name{generate_barcode}
\alias{generate_barcode}
\title{generate_barcode}
\usage{
generate_barcode()
}
\description{
It generates a list of barcodes of 9 bases, based on quaternary Hamming(9,5).
}
\examples{
print("Generating barcodes...")
barcodes=generate_barcode()
print("done.")
print(paste("We just generated", length(barcodes), "barcodes."))

}
